\name{BikeshareData-package}
\alias{BikeshareData-package}
\alias{BikeshareData}
\docType{package}
\title{
Tools for Analyzing Bikeshare Data
}
\description{
  Facilitates analysis of bikeshare data from various municipalities.
}
\details{
\tabular{ll}{
Package: \tab BikeshareData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-01\cr
License: \tab MIT\cr
Depends: \tab methods\cr
}
}
\author{
  Sara Kwasnick and John Sears
  
  Maintainer: Sara Kwasnick and John Sears 
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
